requirements.txt
lxml==4.7.1
python-telegram-bot==13.10
requests==2.26.0
